+++ arch/mips/lib/dump_tlb.c
@@ -112,7 +112,7 @@ void dump_tlb_wired(void)
 void
 dump_tlb_addr(unsigned long addr)
 {
-	unsigned int flags, oldpid;
+	unsigned long flags, oldpid;
 	int index;
 
 	local_irq_save(flags);
